comments pararllelmusic

+++ src/MidiGremlin/Note.cs
@@ -15,8 +15,13 @@ namespace MidiGremlin
         public double Duration { get; set; }
         public byte Velocity { get; set; }
         public int OctaveOffset { get; set; }
-        
 
+        /// <summary>
+        /// Represents a note. A note consist of tone,duration and velocity
+        /// </summary>
+        /// <param name="tone">An enum which represent a tone</param>
+        /// <param name="duration">How long the note is played. In beats</param>
+        /// <param name="velocity">How hard the note is played</param>
         public Note (Tone tone, double duration, byte velocity = 64)
         {
             OctaveOffset = (int)tone/12;
@@ -25,7 +30,13 @@ namespace MidiGremlin
             Velocity = velocity;
 
         }
-
+        /// <summary>
+        /// Offsets the note.
+        /// </summary>
+        /// <param name="offset">Moves by intevals/tonesteps.</param>
+        /// <param name="duration">How long the note is played. In beats</param>
+        /// <param name="velocity">How hard the note is played.</param>
+        /// <returns> Returns a new note where tone is added by offset</returns>
         public Note OffsetBy(int offset, double? duration = null, byte? velocity = null)
         {
             byte tempVelocity = (byte)velocity;
@@ -44,7 +55,14 @@ namespace MidiGremlin
 
             return new Note(Tone + offset, tempDuration, tempVelocity);
         }
-            
+        /// <summary>
+        /// Offsets the note. Overload with scale.
+        /// </summary>
+        /// <param name="scale">The scale you want to offset by</param>
+        /// <param name="offset">Moves by intevals/tonesteps</param>
+        /// <param name="duration">How long the note is played. In beats</param>
+        /// <param name="velocity">How hard the note is played</param>
+        /// <returns>Returns a note offset by a scale</returns>
         public Note OffsetBy(Scale scale, int offset, double? duration = null, byte? velocity = null)
         {
 
@@ -64,7 +82,15 @@ namespace MidiGremlin
 
             return new Note((Tone)(scale.Interval(Tone) + (int)scale[offset]), tempDuration, tempVelocity);
         }
-
+        /// <summary>
+        /// Offsets the note. Overload with octave
+        /// </summary>
+        /// <param name="scale">The scale you want to offset by</param>
+        /// <param name="offset">Moves by intevals/tonesteps</param>
+        /// <param name="OctaveOffset">the octave you want to offset by</param>
+        /// <param name="duration">How long the note is played. In beats</param>
+        /// <param name="velocity">How hard the note is played</param>
+        /// <returns>Returns a note offset by an octave</returns>
         public Note OctaveOffsetBy(Scale scale, int offset, int OctaveOffset, double? duration = null, byte? velocity = null)
         {
 
@@ -86,7 +112,7 @@ namespace MidiGremlin
         }
 
 
-        internal override IEnumerable<SingleBeat> GetChildren (Instrument playedBy, int startTime)
+        internal override IEnumerable<SingleBeat> GetChildren (Instrument playedBy, double startTime)
         {
             byte velocity = Math.Min((byte) 127, Velocity);
